Less noisy file change updates
Rather than send the file change on every TextChanged or TextChangeI
event, start a timer for every change and only flush changes when
things have been stable for 500ms.

This started with only flushing on CursorHold and CursorHoldI but
CursorHoldI seems to interact even worse with plugins that frequently
opn a popup than TextChangedI does, so even when not touching the
keyboard for a long time the changes might not get sent to the server
from insert mode.

Always flush changed before leaving the buffer since this is easier to
do than to try to get back to the buffer content after moving away.
Always (re)stop the timer from FlushFileChanges in case it was called
because of a buffer leave before the timer fired.