Parse messages by content length
Can handle multiple messages arriving together, or receiving only a
partial message.

Whenever a message is received it's appended to a buffer in case it's a
message continuation. Parse out the content length and only consume
that much from the buffer in case multiple messages are buffered.

This is still very fragile to poorly behaving servers - it would not
tolerate a bad content-length.