Call the server for completions suggestions
Add a utility function `compose` to inject a function to translate the
server response into the format expected by the onFound callback.

Upon completion request, flush file change and call the completion
method.

For now only pull out the labels to match the list of strings format
that onFound wants. Eventually this will need to be a more complex
structure to include replacement ranges and extra info.