Add line text to references quickfix items
If the file is open in a buffer, use the bufnr and getbufline in the
quickfix item. If the file is not yet open use the filename and readfile
in the item.

This is a little inefficient if we repeatedly read the same file up to
varying lines, but hopefully disk caching will reduce the impact.