Update tracked files on server start
Closes #7

Ues s:file_versions to know which files have a didOpen sent to the
server. When the server is started all files need to get resent as a
didOpen so that didChange can be used again. Since there are two paths
to trigger a didOpen - already open file when the server starts or a
newly opened buffer - only send didOpen for the files which are not
present in s:file_versions.

- Add lsc#file#trackAll to iterate over buffers of a given filetype and
  send didOpen for each.
- Add lsc#files#clean to mark all files of a filetype as untracked
- Only send didOpen if the file is not already tracked
- Only send didChange if the file is tracked
- Add lsc#file#version to get the file version for the current buffer.
  Could be useful for a status line.
- Add a boolean return from RunCommand so lsc#file#trackAll can be
  called only when the server is actually started.
- Add an optional file_path argument to lsc#util#documentUri since the
  file for didOpen isn't necessarily the current buffer anymore