Allow reuse of ShowHover preview window
Check if a preview window is already open and switch to it instead of
splitting. Avoids `pclose` which can change layouts which were manually
adjusted.

Set `winfixheight` so that further splits with `equalalways` won't make
the preview window bigger than it needs to be - completion preview
windows already set this option.