Switch to lambda and closure syntax
Lambdas are easier to read than the `'v:val'` syntax, and closures are
less clunky that dict functions.

Delete the `lsc#util#compose` helper and inline the the call to the
composed function. Also inline the SearchCompletions function since it's
separation was a leftover from an attempt to support multiple completion
sources.

Make `lsc#diagnostics#convert` function script-private since it was
unused outside the file.