Add LSCServerStatus function
Fixes #32

Replace `s:initialized_servers` and `s:restart_commands` lists with a
map to multiple possible statuses. Initialized servers can be
distinguished between 'starting' and 'running'. Servers which should be
restarted get a status of 'restarting'.

Add a function to find the status for the current server. This could be
useful in a status line, or interactively if the server is misbehaving.