Remove call buffering
Fixes #33

Call buffering was intended to allow didOpen calls to get handled even
when the buffer was opened before the server was ready. This meant that
if buffer was changed between the didOpen call getting buffered and
sent, the wrong content would be sent. Instead, always send didOpen for
all relevant buffers after initialization so that the content is up to
date.

- Remove RunCommand in favor of a single StartByCommand. There is no
  longer a need for a separate method since the didOpen calls are sent
  after initialization, not serve start.
- Remove return value after starting the server.
- Add a return value to `lsc#server#call` indicating whether it was
  ignored. Do not treat ignored `didOpen` calls as having a tracked
  file.
- Remove `s:buffered_calls` variable and all related handling.