First stab at diff algorithm
Towards #3

The basic algorithm is to look at the first character which is
different, and the last character that has a matching postfix, and treat
everything in between as the diff.

Add utility to find start and end of diff from `List<String>` for old
and new buffer content.

Add unit tests based on edge cases that came up during implementation.