Avoid storing empty diagnostic maps
If a large number of empty diagnostics are sent - for example if the
server sends empty dignostics for every file with no errors under some
root - we needlessly fill a map with keys pointing to no useful data. We
already treat a missing key as empty diagnostics so we can stop storing
this info directly.

- If we already have no diagnostics and get an empty list again there is
  nothing to do - bail out.
- If we get an empty list and previously had diagnostics, clear the key
  rather than update it to an empty dict.
- Fix a typo in the double source check.
- Avoid looking up line key again after setting to an empty list.