Fix diagnostic truncatign for multi-byte chars
In some cases where a mult-byte character was at the truncation point it
could get broken and end with with things like `<e2><86>` in the output.
Use `strtrans` to get the string as it would be printed, then `strchars`
and `strdisplaywidth` to truncate to handle both multi-byte and wide
characters.

Add changelog for duplicate word completion and prepare to tag for
release.