Add support for `textDocument/signatureHelp`
This commit adds a command `:LSCLientSignatureHelp` which sends a
`textDocument/signatureHelp` request to the language server. The active
signature is shown in the preview window, and the active parameter is
highlighted using the new group `lscCurrentParameter`.

This commit also introduces the new autocommand `LSCEditPreview` which
is triggered when the content of the preview window is changed. This
allows us to run a `match` on the content of the preview window in order
to highlight the active parameter.