Support completion items with `MarkupContent`
The documentation attached to completion items can be either a string or
a `MarkupContent`. To make sure that a value was actually present, we
used to compare the value of the `documentation` of a completion item
against `v:null`, which would throw an error when the value is present
and is a dictionary (representing the `MarkupContent`).

When a `MarkupContent` is present, we take its `value` and use this as
the completion `info`. We ignore the `kind`.