Use selection to handle exclusive range in edit
Replace a hack with moving the cursor in a different way depending on
where the end of the range falls. vim has built in support for
end-exclusive ranges with the `selection` setting.

Avoid the extra checking and restoring of `paste` within `s:ApplyAll`
since it's handled by the calling method.