Line-wrap long diagnostics and auto-update cmdheight.
Previously, diagnostic messages were processed with `strtrans` and
displayed on a single-line making it difficult to understand long
multi-line diagnostics.

With this commit, we line-wrap long multi-line messages so that they fit
within the editor maximum width and automatically adjusts the
`cmdheight` variable to ensure that there's enough vertical space to
display the message.

We make sure to preserve newlines in the original message even if that
can produce unexpected line wrapping. For example,
```
// Before
// ----------
The quick brown
fox jumps.

// After
// ----------
The quick
brown
fox jumps.

// Expected
// ----------
The quick
brown fox
jumps.
```