Add LSClientLineDiagnostics command (#131)
Fixes #123

When a diagnostic has newlines in the message they get translated into other characters to avoid echoing long messages and triggering a "Press enter to continue" message while moving around in a file. Preserving formatting may be useful for some messages, so add a command to be able to see the message in it's entirety.

When messages are a single line the severity is also echoed in line to avoid taking up too much space, when messages are multiple lines the severity is echoed on its own line to avoid shifting the text since subsequent lines may be intended to line up in particular ways.