Manually expand paths to missing directories
Fixes #138

The best guess for the full path of a relative path buffer in a
non-existent directory is to base it off the `cwd`. This matches the
behavior of the `:p` modifier if the buffer is for a file that does not
exist, but would be in an existing directory.

Buffers that have not been written are not associated with a file. For
these buffers, when the `cwd` changes the apparent full path will also
change. This is true whether using the `:p` modifier, or the new guess
based on `cwd`.

- Add a utility `lsc#file#fullPath()` to incorporate this normalization
  against `cwd` when necessary.
- Update every place using `expand('%:p')` to call this new utility
  instead.