Manually expand paths to missing directories (#139)
Fixes #138

The best guess for the full path of a relative path buffer in a
non-existent directory is to base it off the `cwd`. This matches the
behavior of the `:p` modifier if the buffer is for a file that does not
exist, but would be in an existing directory.

Buffers that have not been written are not associated with a file. For
these buffers, when the `cwd` changes the apparent full path will also
change. This is true whether using the `:p` modifier, or the new guess
based on `cwd`.

Add utility methods it incorporate this normalization against `cwd` when
necessary.

- `lsc#file#fullPath()` replaces `expand('%:p')`.
- `lsc#file#normalize(path)` consistently normalizes buffer names which
  are relative.
- `lsc#file#bufnr(path)` replaces `bufnr(path)` since the normalized
  paths would not correspond to any buffer.

When starting from a path, always use `lsc#file#bufnr` before other
buffer related methods like `getbufline`, `getbufvar`, or `getbufinfo`.