Support multiple contents on hover
What
===
Display in the preview window all the contents items that are returned
by a language server when multiple are returned.

Why
===
Some language servers return multiple contents. Often one will be a
code-block from the source that shows things like function parameter
names, etc, and then another will be the markdown from docs for the item
under the cursor.

At the moment only the first item is taken. In some language servers
multiple items are provided and vim-lsc is truncating the content by
only displaying the first.