Set exit status before calling exit (#134)
If the server process exits very quickly after the `exit` request it's
possible that the `on_exit` handler is triggering before the flush is
finished which would make it appear in the handler that the process
exited before it was called. Set the new status and force the `exit`
request on an apparently not-running server to avoid this bad ordering.