Store capabilities on the server object
Towards #57
Towards #164

Add a capabilities library to filter down and parse out the options this
client cares about. Always store a dictionary of capabilities on the
server objects, and search those when checking for how other features
should behave.

This replaces the previous approach of storing a separate dictionary per
feature to track by filetype. Once we support multiple servers for a
given filetype the single config wouldn't make sense.