Suppress diagnostics on the current line
See #169

When typing it is common for there to be diagnostics while the code is
in an incomplete state. Suppress these diagnostics until after leaving
insert mode.

This has the unfortunate side effect of also removing existing
diagnostics from the current line even if they weren't impacted by the
changed text.