Restore qf idx after upating list
This is close but isn't working quite right yet. Attempts to scan the
list for the next item _after_ whatever was previously selected. If the
previously selected item is still present, it should be chosen. In some
cases this seems to work, but other times the wrong item is getting
chosen...