Don't use a removed channel on exit
Fixes #201

If the server exits after the call to 'shutdown' instead of waiting for
the 'exit' notification then we might have removed the `_channel` for
the server. Check whether the key exists before trying to use it.
Regardless of whether the exit notification is sent, call the `OnExit`
callback to avoid blocking.

Fix the `return` for the 'shutdown' request since otherwise we fail to
track that we were waiting for it.