Fix completion at beginning of the line (#202)
Fixes #183

If none of the completion items had a `textEdit` to find a server
specified start character a heuristic using vim's notion of word
characters is used - in the case of no non-word characters until the
beginning of the line this had a bug which made manual completion use
the cursor position instead.

Add some comments about using 1-based character indexes returned. We use
0-based for `completefunc` use cases, but 1 based for calls to
`complete()`.