Handle diagnostics outside of content (#196)
Fixes #157

For diagnostics which start after the end of a line, highlight the end
of line. For diagnostics which after the last line, highlight the end of
the last line.

Diagnostics which span multiple lines where the start is past end of the
first line are unchanged - the assumption is that the diagnostic
highlighting will still show up on the subsequent lines.