Show diagnostics after the end of a file
Diagnostics with a line after the last line of the buffer are shown with
a trailing highlight on the last line, but it's confusing when the
cursor is there and nothing is echoed. In the case where there is not a
separate diagnostic on the last line, but there is one following it,
echo it. This won't be perfect, we'll still show the current line
diagnostic even if the cursor is closer to the end of the line, but it
will be better than nothing.