Change Neovim popup, make it float above current line (not below)
The existing LSV Vim hover popup will be rendered above the current
cursor item.

Tweak the Neovim floating window to do the same thing. It does look
better this way. Previously it rendered below the current line.

I should have done this from the start, but I am a novice when it comes
to Neovim's floating window fuctionality. Anyway, it is better now.