Allow configuring a custom action menu
Alternative to #231

Change the callback passed to `lsc#edit#findCodeActions` to take an
`OnSelected` argument rather than return the chosen action. This enables
asynchronous menus.

Move the logic for handling multiple formats for a command, or an edit,
into `s:ExecuteCommand`.

Check for a `g:LSC_action_menu` variable and use it instead of the
default action menu if it exists.