Prefer plaintext for hover content
If a server supports a plaintext format it is likely to look nicer since
there isn't support for rendered markdown.

There is no configuration for now to avoid unnecessary noise - if users
ask for the ability to go back to preferring markdown we can add a
variable.

See https://github.com/natebosch/vim-lsc/pull/271