Don't send didSave if capability is missing
Fixes #275

Only set the `sendDidSave` config if the server sends the
`textDocumentSync.save` key, otherwise retain the default which is
`v:false`. Omitting any key, or sending a non-dict, should be considered
as an absent capability.

The `includeText` sub-property of the `SaveOptions` interface is not
supported, so we still only check whether the key is present.