Fix some test flakiness (#281)
Clear out file state even for disabled servers on BufUnload. Without
this when re-enabling the server the `didOpen` message would not be sent
so the test would hang waiting for it.

Use wipeout all buffers with a bang for safety in the test teardown.

Wait for clients to close on test teardown.

Clean up log files after printing them.