Implement line difference search in lua
Towards #273

For very large files with a small edit the diff computation spends most
of it's time verifying the equality of the lines before and after the
difference, lua is much faster to iterate the collections and compare
strings. Add a small function to give a faster path for finding the
index of the first and last line with a difference in lua.