Refactor while loops into for loops
These are more readable, and also a bit more performant. In the case
where there is no difference found the value after the loop will be off
by 1, but this is fine because it was already using `>=` to compare. In
the case of `l:line` the value is not referenced after the loop.