Honor `suppress_stder` during startup
Closes #290

Previously output before initialization was intended to be shown to
reduce the chances of hard to debug issues where the server crashes with
a message that is never seen. Since some servers emit noisy logs even
before initialization we can honor this always. In the future this will
change the default to suppress if the stderr can be logged elsewhere and
retrieved if necessary.