Use a long debounce delay in insert mode
Potential fix for #153

Try to avoid an occasional weird interaction between the timer and
leaving insert mode which seemed to occasionally make the cursor bounce
right and left. A long timer seems to help that, but also is annoying
following edits from normal mode. Use a variable debounce by mode to try
for the best of both.