Handle changing lua base index in vim
Closes #307

neovim and vim lists in lua have a different base index - 1 for neovim
(matches non-vim lua arrays) and 0 for vim. This was updated in vim to
match the others. There was already a handling of the offset difference
between neovim and vim, update this to look for the specific patch which
flipped this in vim.

Eventually the support for older vim can be dropped or checked to gate
usage of lua altogether.