Remove the completion_canceled variable (#314)
The only time this would have any effect is if a character was typed
after the last completion request was made and before the response came
in, but also where the cursor has not moved - for instance if a
character was typed and quickly deleted. This doesn't do anything
useful.