Add a test for edits to the end of the file (#321)
Towards #317

Demonstrates the problem reported. When an edit sets the end position
to be the line following the last line of the file the cursor cannot be
placed in at that position so the edit will be done incorrectly, in this
scenario the edit should be done to the end of the last line.