Reuse location lists instead of appending new ones (#342)
Closes #339

Avoid stomping over location lists used for other purposes.

When there is no location list there is nothing that will be hidden so
it is safe to add a new one for diagnostics, including an empty list if
there are no diagnostics. If there is any existing location list, try
instead to update a previously written list which may either be shown or
hidden on the location list stack. If there was not previously a list
written, or if the previously written list has fallen off the end of the
location list stack, the call to `setloclist` will do nothing. In both
cases this is safe because it is not possible to return to the list or
see it in `:lhistory` so the stale list is never seen. Add a command
`:LSClientWindowDiagnostics` to either resurface an existing list or
write a new one.

- Replace diagnostic and highlight `_version` and `_file` variables with
  `lsc_diagnostics` and `lsc_highlights_source` variables. Use an
  identity check with `is` to avoid the need for multiple variables to
  form a unique ID. Removes the `s:diagnostic_versions` dictionary,
  related function, and updating code paths.
- Rename `updateLocationList` to `UpdateWindowStates` to make it clear
  that other window local state is changed, specifically the
  `w:lsc_diagnostics`  and possibly `w:lsc_location_list_id` variables.
- Split into separate `s:UpdateWindowStates` and `updateCurrentWindow`
  functions to improve clarity. Pull out a private `UpdateWindowState`
  for the shared logic. Split location list updating into more fine
  grained functions as well.
- When writing a new location list store the list ID into a window
  variable so that it can be updated later.
- Add a command `:LSClientWindowDiagnostics` which will either resurface
  a hidden LSC diagnostics list if it is somewhere in `:lhistory`, or
  create a new one. When invoked in a location list window which has a
  location list open associated with a real window, uses the LSC buffer
  associated with that window. This matches the behavior of commands
  like `:lhistory` which may be invoked in either window. Counts how far
  away the LSC diagnostics buffer is and uses `:lolder` or `:lnewer` to
  move there. `:<count>lhistory` does not work in nvim.
- Update docs.