Check server status and unmap keys if not running
When a language server exits for any reason, we want to unmap the
vim-lsc mappings in all of its corresponding buffers (so that we don't
have "dead" mappings lying around). To do this, we can either

1. Iterate through all open buffers and unmap buffer-local mappings
   whenever the server quits
2. Unmap buffer-local mappings on-demand when the buffer is opened, if
   the server is no longer running

Option 2 is the better choice (and is what this commit implements) as
it's faster and leaner (iterating through all open buffers and removing
mappings can be *very* slow if there are a lot of buffers open).