Go to last char if position is beyond last line
Specifically with pyls, if you do a rename, the server seems to return the whole file as a change. Therefore the change starts at position 0/0 (line/character) and ends at last line +1 / 0 (the first character on the line behind the last line of the buffer). This position can't be reached with the cursor.
I added a test for pos beyond last line and modify the position to go to the last character of the last line.
The rename in pyls works this way as expected.