Add trailing newlines to buffer content (#356)
Closes #318

Improves compatibility with some language servers and prevents
diagnostics from appearing for missing newlines.

This is not a full solution, in some cases there really won't be a
trailing newline. Those cases should be rare, and including the newline
always should be an improvement for the vast majority of uses. A more
complete solution would check the settings to see if a trailing newline
would be written, but that appears to retain the server compatibility
issue, or reveal a bug in the diff computation.

```
function! s:eof(bufnr) abort
  let l:has_eol = getbufvar(a:bufnr, '&eol') ||
      \ (getbufvar(a:bufnr, '&fixeol') && !getbufvar(a:bufnr, '&binary'))
  return l:has_eol ? "\n" : ''
endfunction
```