Make the server stub a real class
Allows adding the `initialized` getter to wait for the vim client to
become initialized before further interactions. In some tests this can
prevent sending a request too early, and in some cases prevents moving
on to calling `LSClientDisable` before the server has been initialized,
which causes the client to not call `shutdown` and `exit`. Don't
`listen` on the peer until waiting for the `initialized` getter.