Don't start disabled servers (#370)
Fixes a bug where a disabled server might be started if it was
registered after a buffer is already open with that filetype.

Check whether the server is enabled before starting it. Add a test for
calling `RegisterLanguageServer` with an already open buffer for the
file type. Refactor to a consistent pattern using `get` over `has_key`.