Fix the test
Report stderr for vim.expr (might be helpful in other places too). Fix
the bug surfaced by this output with an extra closing paren.

Send an extra "<cr>" since the multiple sendKeys seem to leave vim in a
state with a "Press Enter to continue", maybe where it prints the
function that was just created?