Don't concatenate as many strings (#409)
String concatenation gets more and more expensive the longer the strings
get. In a period of high throughput messages from the server, where
processing is not keeping up with the rate of message delivery, the cost
of concatenating a longer and longer buffer string becomes overwhelming.

Store the buffer as a list of strings. Only concatenate strings in cases
where a single message spans multiple chunks in the buffer, otherwise
consume messages within each chunk.

There are still some string splicing operations which dominate the
runtime of this method, but the cost is much much smaller than it used
to be. Processing throughput is much higher and bottlenecked on the
request callbacks.

In an extreme cases this makes a drastic difference.

Before:

```
 count  total (s)   self (s)  function
 3173 277.953046 277.656509  <SNR>97_Consume()
```

After:

```
 count  total (s)   self (s)  function
 4927   1.456195   0.985914  <SNR>97_Consume()
```