Optimize LSClientAllDiagnostics updates (#410)
When there are more than 500 diagnostics in the entire workspace they
are not all shown in the quickfix list. Add a
`s:highest_used_diagnostic` variable to remember the position in the
sorted list of file paths where it stopped. Any changes that come in for
later files can skip the update to the quickfix list because they
wouldn't be shown. When a file is removed or reduces the number of
diagnostics but it's diagnostics are in the quickfix list, the value is
dropped because it may be invalid.

To find all diagnostics if there is a valid highest used, pull out only
those files for sorting instead of sorting the entire list. When there
are more than 500 files with diagnostics and no valid highest used
variable, use a quick select algorithm to pick the lowest 500.

Add a 100ms debounce to updates to the quickfix list to reduce the total
number of times `s:AllDiagnostics` is called during periods of high
throughput.

Before:

```
count  total (s)   self (s)  function
12837 7829.367381   0.700760  lsc#diagnostics#setForFile()
```

```
count  total (s)   self (s)
 8011              0.012105   let l:all_diagnostics = []
 8011 7728.788529 226.230393   let l:files = sort(keys(s:file_diagnostics), function('lsc#file#compare'))
```

After:

```
count  total (s)   self (s)  function
60980  68.074209   3.454824  lsc#diagnostics#setForFile()
```

```
count  total (s)   self (s)
   10              0.000040   let l:all_diagnostics = []
   10              0.037433   let l:files = keys(s:file_diagnostics)
   10              0.000073   if exists('s:highest_used_diagnostic')
    3   1.662578   0.046623     call filter(l:files, funcref('<SID>IsUsed', [s:highest_used_diagnostic]))
    7              0.000036   elseif len(l:files) > 500
    1   0.096141   0.000211     let l:files = s:First500(l:files)
   10              0.000011   endif
   10   0.143416   0.004199   call sort(l:files, funcref('lsc#file#compare'))

```