Fix edits in non-current buffers (#441)
A change to better handle edits which claim to be past the end of the
file used the current buffer to determine how long the file is, but the
current buffer is not necessarily the one with edits.

Move the opening of the buffer outside of the built up command so that
it happens before computing the rest of the command and `line('$')` is
accurate.

Move `keepjumps` and `keepalt` to the calling code instead of repeating
them in each function. Drop the second `keepjumps` since the entire
process should be handled.