Set up a repro environment for ltex-ls
Capture some work from debugging #437
Sets up a demo showing that diagnostics are working.

Add a `shell.nix`, `.envrc`, and a derivation to install the current
latest version of `ltex-ls`.

Add a vim config to register the server.

Add a file that has the correct filetype and a sentence that raises a
diagnostic.