Add g:lsc_block_complete_triggers
Closes #444

Allows configuring a list of characters which are ignored if any server
enables them as triggers for autocomplete.

Works around a situation where typing `<enter>` quickly after a trigger
character like `{` does not insert a newline.

Always return to normal mode during teardown in the autocomplete test.
This means that if a test fails in a way where it leaves the editor in
insert mode, like checking an expectation before sending the `<esc>`, it
doesn't time out the overall test.