Replace more uses of col
Use a single utility that wraps charcol when possible, and falls back on
col otherwise. Older vim without charcol will not handle multibyte
characters for now. Continue to use `col` for the tag stack, and
`virtcol` for setting character position. In places where the column is
used to index into a string value, prefer `charcol` when available.